check for ss variance in last 10 iterations
The check to determine whether the species were at a steady state was
comparing the last 5 iterations. In order to make it more sensitive, I
changed this to the last 10 iterations. In other words, the last 10
iterations of a simulation need to be within the given tolerance in
order to be called "at steady state."